[Practice] 13/09

+++ app/Http/Controllers/PostsController.php
@@ -4,13 +4,18 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
 
 class PostsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->only(['create', 'store', 'edit', 'update', 'destroy']);
+    }
+
     /**
-     * Display a listing of the resqource.
+     * Display a listing of the resource.
      *
      * @return \Illuminate\Http\ResponseO
      */
@@ -43,10 +48,6 @@ class PostsController extends Controller
 
         $post = BlogPost::create($validated);
 
-        // BlogPost::create();
-        // $post2 = BlogPost::make();
-        // $post2->save();
-
         $request->session()->flash('status', 'The blog post was created!');
 
         return redirect()->route('posts.show', ['post' => $post->id]);
@@ -73,7 +74,11 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
-        return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
+        $post = BlogPost::findOrFail($id);
+
+        $this->authorize('update-post', $post);
+
+        return view('posts.edit', ['post' => $post]);
     }
 
     /**
@@ -86,6 +91,9 @@ class PostsController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize('update-post', $post);
+
         $validated = $request->validated();
         $post->fill($validated);
         $post->save();
@@ -104,6 +112,9 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize('delete-post', $post);
+
         $post->delete();
 
         session()->flash('status', 'The blog post was deleted!');
